SiteAddressChanger: Rename state dir
